mag_profile/requirements.txt
Flask == 1.0.2
запросов == 2.18.4
dialogflow == 0.4.0
python-dotenv == 0.8.2
pymongo==2.8